Move cached kernel to srt.utils
